v0.4.183: split ManifestValidation.swift into per-concern validators
Phase 4.1 of the audit-driven refactor.

The 817-LOC `ManifestValidation.swift` mixed three independent
validation concerns: dependency-graph cycle detection, pattern-family
schema validation, and notebook-check schema validation.  Every edit
to one concern revalidated the whole file under Swift's type checker.

Split into four files, each under 400 LOC:

  - `ManifestValidation.swift` (72 LOC) — DAG cycle detection only.
    Kept the original filename for the DAG part.
  - `PatternFamilyValidator.swift` (351 LOC) — `validatePatternFamilies`
    and its four private helpers
    (`validatePatternFamilyHeader`, `validatePatternCaseHeader`,
    `validatePatternCaseKindSpecific`, `validateFamilyVariablesAndArgRefs`).
  - `NotebookCheckValidator.swift` (392 LOC) — `validateNotebookChecks`
    plus its `swiftlint:disable cyclomatic_complexity` wrapper.
  - `IdentifierValidation.swift` (36 LOC) — `isValidPythonIdentifier`,
    `isValidIdentifierFragment`, `pythonKeywords`.  Tiny shared
    helpers all three validators reference.

Public API unchanged — `validateManifestDependencies`,
`validatePatternFamilies`, `validateNotebookChecks` still have the
same signatures.

No behaviour change.  1279/1279 tests pass; 0 SwiftLint violations.
